Actualizacion de 90 porciento de endpoint implementados

+++ app/Http/Controllers/Api/PasasteSinTagController.php
@@ -2,31 +2,43 @@
 
 namespace App\Http\Controllers\Api;
 
+use Exception;
 use App\Http\Controllers\Controller;
 use App\Http\Resources\PaseDiarioCollection;
 use App\Http\Resources\PaseDiarioResource;
+use App\Models\Pago;
 use App\Models\PaseDiario;
+use App\Models\PaseDiarioDetalle;
 use Illuminate\Http\Request;
 use GuzzleHttp\Client;
 use Illuminate\Support\Facades\Http;
+use Carbon\Carbon;
+use Illuminate\Support\Facades\DB;
+use PhpParser\Node\Stmt\TryCatch;
 
 class PasasteSinTagController extends Controller
 {
+    private $pst_host = "https://pasastesintag-qa01.i1.cl/api/";
+    private $pst_user = "aconcagua";
+    private $pst_password = "da!9No86nXGJ";
+    private $pst_token = null;
+
     public function login_pst()
     {
         $params_api = [
-            "user" => env('API_PST_USER'),
-            "password" => env('API_PST_PASSWORD')
+            "user" => $this->pst_user,
+            "password" => $this->pst_password
         ];
 
-        $uri = env('API_PST_URL')."login";
+        $uri = $this->pst_host."login";
 
-        $response = Http::post($uri, $params_api);
+        $response = Http::withToken($this->pst_token)->post($uri, $params_api);
         $response = $response->json();
 
         if($response["status"] == "success")
         {
-            return $response;
+            $this->pst_token = $response["data"]["token"];
+            return $response["data"];
         } else{
             return $response;
         }
@@ -34,17 +46,27 @@ class PasasteSinTagController extends Controller
 
     public function refresh_token_pst()
     {
+        if(is_null($this->pst_token))
+        {
+            $this->login_pst();
+        }
+
         $params_api = [
-            "token" => "",
+            "token" => $this->pst_token,
         ];
 
-        $uri = env('API_PST_URL')."refresh";
+        $uri = $this->pst_host."refresh";
 
         $response = Http::post($uri, $params_api);
+        $response = $response->json();
 
-        dd($response->json());
-
-        #Obtener el token
+        if($response["status"] == "success")
+        {
+            $this->pst_token = $response["data"]["token"];
+            return $response["data"];
+        } else{
+            return $response;
+        }
     }
 
     public function show_pase_diario(PaseDiario $pase_diario) : PaseDiarioResource
@@ -54,7 +76,7 @@ class PasasteSinTagController extends Controller
 
     public function index_pases_diarios() : PaseDiarioCollection
     {
-        return PaseDiarioCollection::make(PaseDiario::all());
+        return PaseDiarioCollection::make(PaseDiario::where('Estado',0));
     }
 
     public function ingreso_deuda_pst()
@@ -82,9 +104,206 @@ class PasasteSinTagController extends Controller
 
     }
 
-    #Job planificado cada cierta cantidad de tiempo
-    public function get_transitos_scada()
+    public function get_pases_diarios()//: PaseDiarioResource
     {
+        /* $this->get_transitos_urbanos();
+        $this->get_transitos_interurbanos(); */
+
+        $pases_diarios = DB::table('pase_diario_detalles')
+                ->selectRaw('Identificador,Fecha, Patente, Categoria, COUNT(1) AS CantidadPasesDiarios, SUM(Deuda) AS Deuda, SUM(DeudaTag) AS DeudaTag, 1 AS TipoPatente, 0 AS Estado')
+                ->where('Estado',0)
+                ->groupBy('Identificador', 'Fecha', 'Patente','Categoria')
+                ->get();
+        
+        foreach($pases_diarios as $pase_diario_agrupado)
+        {
+            $pase_diario = PaseDiario::where('Identificador',$pase_diario_agrupado->Identificador )->get();
+
+                if($pase_diario->isEmpty())
+                {
+                    $pase_diario =  new PaseDiario([
+                        "Identificador" => $pase_diario_agrupado->Identificador,
+                        "Patente" => $pase_diario_agrupado->Patente,
+                        "Fecha" => $pase_diario_agrupado->Fecha,
+                        "Categoria" => $pase_diario_agrupado->Categoria,
+                        "Deuda" => $pase_diario_agrupado->Deuda,
+                        "DeudaTag" => $pase_diario_agrupado->DeudaTag,
+                        "TipoPatente" => $pase_diario_agrupado->TipoPatente,
+                        "Estado" => $pase_diario_agrupado->Estado
+                        ]);
+                       
+                    $pase_diario->save();
+
+                    $affected = DB::table('pase_diario_detalles')
+                                ->where('Identificador', $pase_diario->Identificador)
+                                ->update(['pase_diario_id' => $pase_diario->id]);
+                }
+        }
+
+        # Envío de deuda a PasasteSinTag.cl
+        #Desarrollar solicitud
+        return PaseDiario::where('Estado',0);
+
+    }
+
+    #Job planificado cada cierta cantidad de tiempo
+    private function get_transitos_urbanos()
+    {
+        $date = Carbon::now()->subDays(1)->format('Ymd');
+        $date_tomorrow = Carbon::tomorrow()->subDays(1)->format('Ymd');
+
+        $query_pd_urbano = "SELECT
+            '00'+CONVERT(VARCHAR, t.FechaHora, 112)+'12'+RTRIM(LTRIM(dt.Patente)) AS 'Identificador'
+            , RTRIM(LTRIM(dt.Patente)) AS 'Patente'
+            , CONVERT(VARCHAR, t.FechaHora, 23) AS 'Fecha'
+            , dt.ClaseCategoria
+            , cc.Descripcion AS 'ClaseCategoriaDescripcion'
+            , dt.Categoria
+            , c.Descripcion AS 'CategoriaDescripcion'
+            , pdf.ImporteRedondeado AS 'Deuda'
+	        , SUM(t.Importe)/100 AS 'DeudaTag'
+            , MIN(t.NumCorrCA) AS 'NumCorrCA'
+        FROM Transitos t (NOLOCK)
+            INNER JOIN DetalleTransitos dt (NOLOCK) ON dt.IDDetalleTransito = t.IDDetalleTransito
+            INNER JOIN CategoriasClases cc (NOLOCK) ON cc.ID_CategoriasClases = dt.ClaseCategoria
+            INNER JOIN Categorias c (NOLOCK) ON c.CodigoCategoria = dt.Categoria AND c.IDCategoriaClase = dt.ClaseCategoria
+            INNER JOIN PasesDiariosTarifas pdf (NOLOCK) ON pdf.CodigoCategoria = dt.Categoria AND pdf.IDCategoriaClase = dt.ClaseCategoria AND pdf.AnioVigente = YEAR(t.FechaHoraCreacion)
+        WHERE
+            t.Estado = 6 
+            AND t.NumeroPuntoCobro IN (1,2,3,4,5,6,7,8)
+            AND t.FechaHora BETWEEN '$date' AND '$date_tomorrow' 
+        GROUP BY 
+            '00'+CONVERT(VARCHAR, t.FechaHora, 112)+'12'+RTRIM(LTRIM(dt.Patente))
+            , dt.Patente  
+            , CONVERT(VARCHAR, t.FechaHora, 23)
+            , dt.ClaseCategoria 
+            , cc.Descripcion 
+            , dt.Categoria 
+            , c.Descripcion 
+            , pdf.ImporteRedondeado 
+        HAVING
+            Patente <> '' AND ((LEN(Patente) IN (5) AND Categoria = 4) OR (LEN(Patente) IN (6))) 
+            ORDER BY dt.Patente ASC, CONVERT(VARCHAR, t.FechaHora, 23) ASC, dt.Categoria DESC";
+        
+        try
+        {
+            $pases_diarios_urbanos = DB::connection('sqlsrv')->select($query_pd_urbano);
+            $pases_diarios_urbanos = collect($pases_diarios_urbanos)->unique(['Identificador']);
+            
+            foreach($pases_diarios_urbanos as $pase_diario_urbano)
+            {
+                $pase_diario = PaseDiarioDetalle::where('NumCorrCA',$pase_diario_urbano->NumCorrCA )->get();
+
+                if($pase_diario->isEmpty())
+                {
+                    $pase_diario =  new PaseDiarioDetalle([
+                        "Identificador" => $pase_diario_urbano->Identificador,
+                        "Patente" => $pase_diario_urbano->Patente,
+                        "Fecha" => $pase_diario_urbano->Fecha,
+                        "ClaseCategoria" => $pase_diario_urbano->ClaseCategoria,
+                        "ClaseCategoriaDescripcion" => $pase_diario_urbano->ClaseCategoriaDescripcion,
+                        "Categoria" => $pase_diario_urbano->Categoria,
+                        "CategoriaDescripcion" => $pase_diario_urbano->CategoriaDescripcion,
+                        "Deuda" => $pase_diario_urbano->Deuda,
+                        "DeudaTag" => $pase_diario_urbano->DeudaTag,
+                        "NumCorrCA" => $pase_diario_urbano->NumCorrCA,
+                        "Estado" => 0
+                        ]);
+                       
+                        $pase_diario->save();
+                }
+            }
+            return 'success';
+
+        } catch (Exception $e)
+        {
+            return 'Hay un error';
+        }
+    }
+
+    private function get_transitos_interurbanos()
+    {
+        $date = Carbon::now()->subDays(1)->format('Ymd');
+        $date_tomorrow = Carbon::tomorrow()->subDays(1)->format('Ymd');
+
+        $query_pd_interurbano = "SELECT
+            '00'+CONVERT(VARCHAR, t.FechaHora, 112)+'12'+RTRIM(LTRIM(dt.Patente)) AS 'Identificador'
+            , RTRIM(LTRIM(dt.Patente)) AS 'Patente' 
+            , CONVERT(VARCHAR, t.FechaHora, 23) AS 'Fecha'
+            , dt.ClaseCategoria 
+            , cc.Descripcion AS 'ClaseCategoriaDescripcion' 
+            , CASE 
+                WHEN dt.Categoria = 1 THEN 4 
+                WHEN dt.Categoria = 2 THEN 1 
+                WHEN dt.Categoria = 3 THEN 1 
+                WHEN dt.Categoria = 4 THEN 2 
+                WHEN dt.Categoria = 5 THEN 2 
+                WHEN dt.Categoria = 6 THEN 3 
+                WHEN dt.Categoria = 7 THEN 3 
+                ELSE 1 
+            END AS 'Categoria' 
+            , c.Descripcion AS 'CategoriaDescripcion' 
+            , pdf.ImporteRedondeado AS 'Deuda'
+	        , SUM(t.Importe)/100 AS 'DeudaTag'
+            , MIN(t.NumCorrCA) AS 'NumCorrCA'
+        FROM Transitos t (NOLOCK) 
+            INNER JOIN DetalleTransitos dt (NOLOCK) ON dt.IDDetalleTransito = t.IDDetalleTransito 
+            INNER JOIN CategoriasClases cc (NOLOCK) ON cc.ID_CategoriasClases = dt.ClaseCategoria 
+            INNER JOIN Categorias c (NOLOCK) ON c.CodigoCategoria = dt.Categoria AND c.IDCategoriaClase = dt.ClaseCategoria 
+            INNER JOIN PasesDiariosTarifas pdf (NOLOCK) ON pdf.CodigoCategoria = dt.Categoria AND pdf.IDCategoriaClase = dt.ClaseCategoria AND pdf.AnioVigente = YEAR(t.FechaHoraCreacion)
+        WHERE 
+            t.Estado = 6 
+            AND t.NumeroPuntoCobro IN (9,10)
+            AND t.FechaHora BETWEEN '$date' AND '$date_tomorrow'
+        GROUP BY
+            '00'+CONVERT(VARCHAR, t.FechaHora, 112)+'12'+RTRIM(LTRIM(dt.Patente))
+            , dt.Patente  
+            , CONVERT(VARCHAR, t.FechaHora, 23) 
+            ,t.FechaHora
+            , dt.ClaseCategoria
+            , cc.Descripcion
+            , dt.Categoria
+            , c.Descripcion
+            , pdf.ImporteRedondeado 
+        HAVING
+            dt.Patente <> ''
+            AND ((LEN(dt.Patente) IN (5) AND Categoria = 4) OR (LEN(dt.Patente) IN (6)))
+        ORDER BY dt.Patente ASC, CONVERT(VARCHAR, t.FechaHora, 23) ASC, dt.Categoria DESC";
+
+     try {
+
+        $pases_diarios_interurbanos = DB::connection('sqlsrv')->select($query_pd_interurbano);
+        $pases_diarios_interurbanos = collect($pases_diarios_interurbanos);
+
+        foreach($pases_diarios_interurbanos as $pase_diario_interurbano)
+            {
+                $pase_diario = PaseDiarioDetalle::where('NumCorrCA',$pase_diario_interurbano->NumCorrCA )->get();
+
+                if($pase_diario->isEmpty())
+                {
+                    $pase_diario =  new PaseDiarioDetalle([
+                        "Identificador" => $pase_diario_interurbano->Identificador,
+                        "Patente" => $pase_diario_interurbano->Patente,
+                        "Fecha" => $pase_diario_interurbano->Fecha,
+                        "ClaseCategoria" => $pase_diario_interurbano->ClaseCategoria,
+                        "ClaseCategoriaDescripcion" => $pase_diario_interurbano->ClaseCategoriaDescripcion,
+                        "Categoria" => $pase_diario_interurbano->Categoria,
+                        "CategoriaDescripcion" => $pase_diario_interurbano->CategoriaDescripcion,
+                        "Deuda" => $pase_diario_interurbano->Deuda,
+                        "DeudaTag" => $pase_diario_interurbano->DeudaTag,
+                        "NumCorrCA" => $pase_diario_interurbano->NumCorrCA,
+                        "Estado" => 0
+                        ]);
+                       
+                        $pase_diario->save();
+                }
+            }
+
+            return 'success';
+
+     } catch (Exception $e) {
+        return 'Hay un error';
+     }
 
     }
 
@@ -135,19 +354,76 @@ class PasasteSinTagController extends Controller
   
     }
 
-    public function eliminar_registro_deuda_pasaste_sin_tag()
+    public function eliminar_registro_deuda_pasaste_sin_tag(Request $request)
     {
-        $uri = env('API_PST_URL')."eliminar_deuda_id";
+        $uri = $this->pst_host."eliminar_deuda_id";
 
-        $params_api = [
-            "identificador" => [
-                "","",""
-            ]
-        ];
+        $pase_diario = PaseDiario::where('Identificador',$request->identificador);
+        $pase_diario->Estado = 99; # Solicitar eliminación de deuda
+        $pase_diario->save();
 
-        $response = Http::delete($uri, $params_api);
+        if(!$pase_diario->isEmpty())
+        {
+            $params_api = [
+                "identificador" => [$request->identificador]
+            ];
+    
+            $response = Http::delete($uri, $params_api);
 
-        dd($response->json());
-   
+            if($response->successful())
+            {
+                $pase_diario->Estado = 100; # Informar que deuda ya fue eliminada de pasastesintag.cl
+                $pase_diario->save();
+                return response()->json($data = [
+                    'mensaje' => 'Deuda eliminada con exito'
+                ], $status = 200);
+            } else {
+                return response()->json($data = [
+                    'mensaje' => 'Deuda no pudo ser eliminada de pasastesintag.cl'
+                ], $status = 404);
+            }
+
+        } else {
+            return response()->json($data = [
+                'mensaje' => 'No existe deuda en nuestro sistema acerca de este identificador: '.$request->identificador
+            ], $status = 404); 
+        }  
     }
+
+    public function anular_pago(Request $request)
+    {
+        $uri = $this->pst_host."anular_pago";
+
+        $pago = Pago::where('Comprobante',$request->comprobante);
+        $pago->Estado = 80; # Solicitar anulación de pago en pasastesintag, debido a que Recaudador no informo pago en rendicion
+        $pago->save();
+
+        if(!$pago->isEmpty())
+        {
+            $params_api = [
+                "comprobante" => [$request->comprobante]
+            ];
+    
+            $response = Http::post($uri, $params_api);
+
+            if($response->successful())
+            {
+                $pago->Estado = 81; # Pago anulado en pasastesintag, pasediario disponible para volver a cobrar
+                $pago->save();
+                return response()->json($data = [
+                    'mensaje' => 'Comprobante de pago anulado con exito'
+                ], $status = 200);
+            } else {
+                return response()->json($data = [
+                    'mensaje' => 'Comprobante de pago no pudo ser anulado en pasastesintag.cl'
+                ], $status = 404);
+            }
+
+        } else {
+            return response()->json($data = [
+                'mensaje' => 'No existe comprobante de pago en nuestro sistema acerca de este comprobante: '.$request->comprobante
+            ], $status = 404); 
+        }  
+    }
+
 }
